requirements.txt
beautifulsoup4==4.12.0
Django==4.2
requests==2.28.2
tabula-py
tabulate==0.9.0
django-crispy-forms
lxml